changes to ui and file structure
Updated the ui to be more responsive to smaller screens. There is still
a bug where the rules box causes the page to become larger than the
window if the window is shrunk but only a little. This is because the
bootstrap container that holds the rules is adjusting to screen size by
making the rules container thinner and taller, but this messes things
up. Will need to look into it.

The amount of code in the app.js file was getting to much to handle it
apropriatly, so I seperated the code into three parts. The rendering, the
main game code, and the web stuff (eventhough that hasn't been written
yet). It's a lot better already, but it could probably do with a little
more cleanup within the individual files.