create online game working
fixed the problems with creating an online game. Data is now sent and
recieved properly, but there is still a lot of work needed in the ui
department. The TODO's in the code should just about cover what needs to
be done(for the ui), but don't be suprised if they don't.